Oilpan: Reapply: Remove all the RefPtrs and most of the raw pointers to style sheets.
Comment on the cases where the raw pointers cannot be removed at this point.

Leave the raw pointers in inspector/ alove. They are safe as they are
now because the inspector leaks all the CSSStyleSheets it is inspecting
via Persistent handles at this point (and used to leak them all via
RefPtrs.)

Review URL: https://codereview.chromium.org/196513007

git-svn-id: svn://svn.chromium.org/blink/trunk@169462 bbb929c8-8fbe-4397-9dbb-9b2b20218538