Fix an issue that fullscreen layer is scrollable
When an element enters fullscreen, main frame's content size may
not be equal to that of the fullscreen renderer. As a result, the
fullscreen layer is scrollable. This might cause some unnecessary
computations and some wierd artifacts:

For example, on android phones, the edge glow effect may come after
scrolling the fullscreen video page several times, rather than coming
immediately.

This CL sets the scroll clip layer to NULL, so we won't be able to scroll it.

BUG=352183

Review URL: https://codereview.chromium.org/200943002

git-svn-id: svn://svn.chromium.org/blink/trunk@169566 bbb929c8-8fbe-4397-9dbb-9b2b20218538