Explicitly mark first 2 args of addEventListener/removeEventListener as optional
addEventListener and removeEventListener should require 2 arguments
(per spec), but legacy content calls these with 0 or 1 arguments.

For background, see this bug:
https://code.google.com/p/chromium/issues/detail?id=249598

This results in a special case in the CG, and an IDL that doesn't reflect
reality.

This updates the IDL to reflect reality, and removes the special-casing
from the CG, removing 3 FIXMEs!

There's now only 1 EventTarget FIXME in the templates, namely
the frame security check (on window).

This changes
addEventListener.length
removeEventListener.length
...from 2 to 0, reflecting reality.

R=haraken
BUG=345503
BUG=249598

Review URL: https://codereview.chromium.org/203603005

git-svn-id: svn://svn.chromium.org/blink/trunk@169631 bbb929c8-8fbe-4397-9dbb-9b2b20218538