Web Animations API: Update interpolation test harness to test element.animate()
The animate() method on Element is severely under tested. This patch makes
use of the existing animation interpolation tests to also test element.animate().
Just over 50% of the tests pass, the failures are being checked in as expected
failures for now as it is representative of the current state of element.animate().

BUG=352472

Review URL: https://codereview.chromium.org/203723005

git-svn-id: svn://svn.chromium.org/blink/trunk@169775 bbb929c8-8fbe-4397-9dbb-9b2b20218538