[CSS Grid Layout] Prevent issues with checkLayout() in grid items
When we use checkLayout() on a grid item, the results (PASS or FAIL
messages) are added as new auto-placed items in the grid. This causes
that the grid is marked as dirty which could hide bugs in the tests.

Modify the related layout tests to use a different container for the
test results. Updated expectations accordingly.

There is one test missing
"grid-item-addition-auto-placement-update.html". Which will be modified
in the patch to fix the bug in RenderGrid::addChild() (see
https://codereview.chromium.org/198703004/).

Review URL: https://codereview.chromium.org/208133003

git-svn-id: svn://svn.chromium.org/blink/trunk@169777 bbb929c8-8fbe-4397-9dbb-9b2b20218538