Use |isolate| local variable in bindings, clean up comments
This removes 6 FIXMEs, and removes duplicate calls of info.GetIsolate().

This updates the license block to the new form (since this is *generated* code),
and makes it automatically use the correct module name.
I didn't make the year update automatically, as otherwise r-b-t will break
every new year and this detail doesn't seem worth the hassle.

I also moved one long comment into a template comment, so it's not in the
generated code. This seems clearer (it's in the template), and reduces
duplication and noise.

Overall these reduce the size of the generated code and tests
by about 6% (generated code from 17 MB from 16 MB)!
Not a big deal, but makes them a bit faster to compile and easier to read.

R=haraken
BUG=345503

Review URL: https://codereview.chromium.org/210663003

git-svn-id: svn://svn.chromium.org/blink/trunk@169910 bbb929c8-8fbe-4397-9dbb-9b2b20218538