Add an overloaded version of findPlainText() accepting two Positions.
This patch is in continuation of <https://codereview.chromium.org/210363006>.
While the last patch added an overloaded TextIterator constructor, this patch
adds similar overloads to findPlainText(), which will be used in the find-in-
page implementation in WebFrameImpl.

The added function is not used yet. Hence this patch should have no impact on
the product behavior.

BUG=353430

Review URL: https://codereview.chromium.org/203973005

git-svn-id: svn://svn.chromium.org/blink/trunk@170025 bbb929c8-8fbe-4397-9dbb-9b2b20218538