Schedule image resource downloads by decreasing visual impact
This CL adds an intra-priority level sorting value for finer grained
control of resource loading. The blink side of this patch sorts images by
estimated on screen real-estate, preferring to download images in order of
decreasing visual impact. Ie. On a page with a large banner, and dozens of
small icons, the large banner will most likely be scheduled ahead of the
others.

In the future, we can also use this to experiment with other signals, ie.
prioritize out of viewport images by their distance to viewport.

Will be followed up by https://codereview.chromium.org/146333004/.

In our tests, this was a nice win on speed-index for desktop, with a gain of 
about ~2-3%. 

This is first of 3 sided patch. Following pattern suggested on chromium-dev:

"
1) Add the new WebKit interface, preserving the old.
2) Change Chrome to use the new interface.
3) Modify WebKit to remove the old interface.
"

Review URL: https://codereview.chromium.org/153983004

git-svn-id: svn://svn.chromium.org/blink/trunk@170072 bbb929c8-8fbe-4397-9dbb-9b2b20218538