Protecting against other possible uses of invalid V8 execution context.
Speculative fix for a user reported crash when a worker with an active IndexedDB
transaction was terminated. This stack is very similar to the one for 345354. I
do not have a reproduction for this bug was unable to verify that it fixes the
problem, but checking for a stopped context is still a good idea.

BUG=356146, 356171

Review URL: https://codereview.chromium.org/213073004

git-svn-id: svn://svn.chromium.org/blink/trunk@170107 bbb929c8-8fbe-4397-9dbb-9b2b20218538