Don't inherit block style if text is inserted after paragraph selection.
It is reported that, under some condition, an unexpected <div> is inserted
when text is inserted after a paragraph is selected. This was caused by
InsertText::doApply() adjusting the text style according to the current typing
style after the insertion.

This patch clears block properties from the typing style if the selection ends
at the start of a block.

Additionally, this patch fixes another issue in DeleteSelectionCommand, which
may erroneously insert a placeholder <br> if the return value of
VisiblePosition::previous() or next() is a null VisiblePosition.

BUG=349531

Review URL: https://codereview.chromium.org/214313002

git-svn-id: svn://svn.chromium.org/blink/trunk@170157 bbb929c8-8fbe-4397-9dbb-9b2b20218538