De-bloat SVGPathStringBuilder
For a class this meager, it was a bit of a heavy-weight:

SVGPathStringBuilder::arcTo() clocked in at a whopping 7891 bytes
(Linux, x86-64, gcc-4.8.2), but even the much simpler lineToVertical()
used > .5k of text-space.
Sacrifice the StringOperators-based pattern and use a cascade of functions
instead.
Besides reducing footprint, this also eliminates some of the copying that
would occur between various temporary String objects.
With the change in place, arcTo() dropped to 640 bytes (and lineToVertical
got cut in half.)

Review URL: https://codereview.chromium.org/212603003

git-svn-id: svn://svn.chromium.org/blink/trunk@170523 bbb929c8-8fbe-4397-9dbb-9b2b20218538